requirements.txt
# Core Agno dependencies
agno>=2.0.0
fastapi[standard]>=0.115.0
uvicorn>=0.32.0

# LLM providers
openai>=1.0.0
anthropic>=0.40.0

# MCP support
mcp>=1.0.0

# Database
sqlalchemy>=2.0.0

# Scheduling for background tasks
apscheduler>=3.10.0

# Environment variables
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0
httpx>=0.27.0
fastmcp

